added parameters validation for add/edit domain added defaultChecked for active checkmark in addDomain changed disabled fields css to display: none
