+++ src/components/addDomainModal/addDomainModal.jsx
@@ -1,8 +1,8 @@
-import React, { Component, useState } from 'react';
+import React, {Component, useEffect, useState} from 'react';
 import Modal from 'react-bootstrap/Modal'
 import Button from "react-bootstrap/Button";
 import './addDomainModal.scss';
-
+import {validateParameters} from "../../common";
 
 import { useSelector, useDispatch } from "react-redux";
 import { NotificationManager } from 'react-notifications';
@@ -26,39 +26,6 @@ function AddDomainModal(props) {
 }
 
 function DomainModal(props) {
-
-    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-    const [validJSONorXML, setValidJSONorXML] = useState(true);
-
-    const checkValidJSONorXML = function check() {
-        let parameters = document.querySelector("textArea[name=\"parameters\"]").value;
-        if (getSelectedServiceType == 0) {
-            setValidJSONorXML(IsValidJSON(parameters));
-        }
-        else {
-            setValidJSONorXML(IsValidXML(parameters));
-        }
-    }
-
-    function IsValidJSON(str) {
-        try {
-            JSON.parse(str);
-        } catch (e) {
-            return false;
-        }
-        return true;
-    }
-
-    function IsValidXML(str) {
-        let parser = new DOMParser();
-        let xml = parser.parseFromString(str, "application/xml");
-        let isValid = xml.querySelector("parsererror") == null ? true : false;
-        return isValid;
-    }
-    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-    ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
 
     const isLogged = useSelector(state => state.isLogged);
     const token = useSelector(state => state.token);
@@ -170,7 +137,7 @@ function DomainModal(props) {
         event.preventDefault();
     }
 
-
+    useEffect(()=>{validateParameters();}, [getSelectedServiceType]); //validates parameters on serviceType change
 
     return (
         <>
@@ -203,8 +170,7 @@ function DomainModal(props) {
                         <input className="BasicAuthDisable" type="text" placeholder="User" name="user" disabled={isUsernamePasswordDisabled()} required max="1024" />
                         <input className="BasicAuthDisable" type="password" placeholder="Password" name="password" disabled={isUsernamePasswordDisabled()} required max="1024" />
                         <textarea className="textArea" form="formForPost" rows="4" name="parameters" placeholder="Parameters" disabled={isParametersDisabled()} required max="4096"
-                            onBlur={checkValidJSONorXML}></textarea>
-                        {validJSONorXML ? "" : "Invalid JSON/XML"}
+                            onChange={validateParameters}></textarea>
                         <input className="SelectInterval" type="number" placeholder="Interval" name="interval" min="3" max="2000000" required />
                         <input className="SelectIntervalSeconds" disabled="disabled" type="text" placeholder="  (s)" />
                         <input className="SelectInterval" type="number" placeholder="Amber threshold" name="threshold" min="10" max="60000" required />
@@ -212,12 +178,12 @@ function DomainModal(props) {
 
                         <div className="float-left d-flex inlineItems">
                             <label htmlFor="checkboxTitle2" className="d-inline-block text-left aLabel">Active: </label>
-                            <input className="d-inline-block aCheckbox float-left" id="checkboxTitle2" type="checkbox" name="active" value="active"></input>
+                            <input className="d-inline-block aCheckbox float-left" id="checkboxTitle2" type="checkbox" name="active" defaultChecked></input>
                         </div>
 
                         {/* <button>Test(sitas dar neveikia)</button> */}
                         <div className="d-flex flex-row inlineItems">
-                            <button type="submit" value="send POST" disabled={!validJSONorXML}>Add</button>
+                            <button type="submit" value="send POST">Add</button>
                             <button variant="primary" onClick={handleClose}>Cancel</button>
                             <button onClick={testService}>Test</button>
                         </div>
