+++ src/components/editDomainModal/editDomainModal.jsx
@@ -3,6 +3,7 @@ import Modal from 'react-bootstrap/Modal'
 import Button from "react-bootstrap/Button";
 import './editDomainModal.scss';
 import DeleteDomain from "../deleteDomain/deleteDomain";
+import {validateParameters} from "../../common";
 
 import Icon from './../../Content/edit_icon.png';
 
@@ -154,7 +155,7 @@ function EditDomain(props) {
     }
     const [response, setResponse] = useState(); //response from server
 
-    //useEffect(validateJsonXML(), [getIsRestSoapSelected]);
+    useEffect(()=>{validateParameters();}, [getIsRestSoapSelected]); //validates parameters on serviceType change
 
     return (
         <>
@@ -187,7 +188,7 @@ function EditDomain(props) {
                         <input className="BasicAuthDisable" name="password" defaultValue={props.domain.auth_Password} disabled={isUsernamePasswordDisabled()} type="password"
                             placeholder="Password" required max="1024" />
                         <textarea className="BasicAuthDisable" name="parameters" defaultValue={props.domain.parameters} className="textArea" rows="4"
-                            placeholder="Parameters" disabled={isParametersDisabled()} required max="4096" onChange={validateJsonXML}></textarea>
+                            placeholder="Parameters" disabled={isParametersDisabled()} required max="4096" onChange={validateParameters}></textarea>
                         <input name="interval" defaultValue={Math.trunc(props.domain.interval_Ms / 1000)} type="number"
                             placeholder="Interval" required min="3" max="2000000" />
                         <input className="SelectInterval" defaultValue={props.domain.latency_Threshold_Ms} type="number" placeholder="Amber threshold" name="threshold"
@@ -215,40 +216,6 @@ function EditDomain(props) {
             </Modal>
         </>
     );
-
-    function validateJsonXML() {
-        const body = document.querySelector("body");
-        console.log("body class: " + body.className);
-        if (body.className == "modal-open") {
-            const parametersTextArea = document.querySelector("textarea[name=\"parameters\"]");
-            let parameters = parametersTextArea.value;
-            console.log("getIsRestSoapSelected: " + getIsRestSoapSelected);
-            if (getIsRestSoapSelected == 0 && !IsValidJSON(parameters)) {
-                parametersTextArea.setCustomValidity("Invalid JSON");
-            } else if (getIsRestSoapSelected == 1 && !IsValidXML(parameters)) {
-                parametersTextArea.setCustomValidity("Invalid XML");
-            } else {
-                parametersTextArea.setCustomValidity("");
-            }
-            console.log("valid: " + parametersTextArea.checkValidity());
-        }
-    }
-
-    function IsValidJSON(str) {
-        try {
-            JSON.parse(str);
-        } catch (e) {
-            return false;
-        }
-        return true;
-    }
-
-    function IsValidXML(str) {
-        let parser = new DOMParser();
-        let xml = parser.parseFromString(str, "application/xml");
-        let isValid = xml.querySelector("parsererror") == null ? true : false;
-        return isValid;
-    }
 
     function handleSubmit(event) {
         setResponse("waiting...");
