+++ src/components/editDomainModal/editDomainModal.scss
@@ -167,6 +167,3 @@
     padding: 0 !important;
   }
 }
-.BasicAuthDisable:disabled{
-  background : #e1e1e1;
-}
